WSPingPong C# webservice and client
1. The folder WSPingPongCS contains a self-hosted C# web service
implemented without interface. A new method bool ball_Check() is added
to check that wether the service has a ball already. This method should
be called by your own client before sending a ball to the service when
you consider it as necessary too.
2. The folder WSPingPongCSClient contains a C# client in which 2
references for accessing the 2 web services are configured. The 2
references should be reconfigured in your own test environment.